test: handle errors in server API tests
The `test:api` script for the server would previously leave the test server
running if any tests failed. This makes sure the server gets stopped even if the
tests failed, and still ensures the correct exit code gets surfaced so the whole
test script shows up correctly as an error if it failed.

This also makes sure that misconfiguration of the test is an an actual error,
rather than printing an error message but then exiting successfully.