test: more error handling corrections
It turns out there was a global teardown for the API tests that would cause
Mocha to always exit successfully, even if there was a failed test. I've
corrected this by doing what I *think* this was trying to handle: shutting
down the global database connection(s) created in `utils.js` and that are
otherwise left hanging and keep the test process running.

This also simplifies `resetDb()` by allowing it to just throw an error, since
the only place it's called from just manually throws the error that it caught.