Update the db seed files with expected default values.
Assuming that the entire app is stood up from scratch,
there should be a pretty minimal set of tenants, users, and agencies by default.
Additional tenants, agencies, and users should be populated by the user instead.

This change deletes a lot of legacy values left over from when the app was intended
to be deployed in a distributed manner.