fix: standardize client-side vscode debugging (783) (#784)
There are no current .vscode files checked into the repo to aid
in debugging client-side code. It would be helpful to have these
in order to better standardize debugging procedures, at least for
those running VSCode.

This is a simple first pass. To use it:
1. Open usdr-gost.code-workspace
2. Go to the "Run and Debug" menu
3. Select "Debug in Chrome (Client)". Under the hood this
will run a dev server, and open a debug copy of Chrome.
4. Set breakpoints in client-side code, see that they're hit!

References:
1. [Debugging in VSCode](https://v2.vuejs.org/v2/cookbook/debugging-in-vscode.html)
2. [microsoft/vscode-recipes](
https://github.com/microsoft/vscode-recipes/tree/master/vuejs-cli)

Co-authored-by: Nat Hillard <nat.hillard@gmail.com>