fix: #1268 - add Invalidate button to the UI
This fix adds an invalidate option to the UI for uploads. The Uploads
page now has additional logic to show when something has been manually
invalidated. A new download button was created. It looks identical to
the smaller in-line version, but is a bigger button. Test cases have
also been added.
A few pieces have been added to arpa reporting to show icons.