feat: enable saved search digest emails per user
Introduces an endpoint to enable admin users to send digest emails for
all saved searches belonging to a particular user. This is a precursor
to sending saved search in bulk to all users as part of the cron. This
change also includes some adjustments to share email digest methods with
the previous iteration.