Replace bootstrap-vue icons with vue3-bootstrap-icons library
The bootstrap-vue library somewhat supports vue 3 with the compatibility
layer for vue 2 syntax, but their icons definitely do not work. (They
rely on `Vue.extend` to set up the icons, which is no longer valid syntax.)
I've found this library to work pretty well as a drop-in replacement, though
some syntax and styling might be slightly different.